split events member/rush into two separate services/APIs
